[TNT-245] fix: dietId 데이터 타입 수정

+++ core/navigation/src/main/java/co/kr/tnt/navigation/RouteModel.kt
@@ -105,7 +105,7 @@ sealed interface Route {
     data object TraineeMealRecord : Route
 
     @Serializable
-    data class TraineeMealRecordDetail(val id: Int) : Route
+    data class TraineeMealRecordDetail(val id: Long) : Route
 
     @Serializable
     data class WebView(val url: String) : Route
